feat: add lazy emoji picker hook
